Making things more readable and simple.
